Remove hotfix for cypress fail on initial hit
There was a bug where on initial run in dev the site would show a spinner until
refreshed at least once. Mitigated this by adding an extra .visit() to
cypress, essentially forcing it to refresh.

Not a big deal, but pointless if the bug is gone. Let's see - this might
still be required in the CI